Autogenerate both the performance overview and year listings.
The comment tells the story. I'm probably leaning too hard into the macrology
but it makes it much simpler to start an individual day and know it will be
included in the site build. I find it slightly awkward that the year sections
are still manually listed in `@ADVENT` considering they are autogenerated now.
But making the macrology even more clever just didn't seem worth it and I have a
note about how to handle it in the README.md. If I ever change my mind,
something like this would automate it:

```
(defmacro generate-docs ()
  (flet ((generate-year-section-locatives ()
           (let* ((src-dir (asdf:system-relative-pathname :advent "src/"))
                  (subdirs (uiop:subdirectories src-dir)))
             (loop for subdir in (reverse subdirs)
                   for year = (parse-integer (lastcar (pathname-directory subdir)))
                   for symbol = (symbolicate "@AOC." (write-to-string year))
                   collecting `(,symbol section)))))
    `(defsection @advent (:title "Advent of Code")
       (@links section)
       (@background section)
       ,@(generate-year-section-locatives)
       (@overview section)
       (@aoc.util section))))

(generate-docs)
```