Find an even better way to factor Day 2.
Of course, we can do even better than accepting a strategy option.
Since there are only 9 possible RPS games that can be played, simply
parse the unique attributes of the game based on the input data.
Now the code for playing and scoring the games is completely uniform.
The only thing that varies is how we interpret the input and we select
the game that has all of the unique attributes specified by that input.