Don't funcall things before they exist.
We may have a bit of a problem here in that the AOT compiled
DEFSUMMARY forms can't benchmark the PART-1 and PART-2 functions
that don't exist yet. An extremely gross but pragmatic solution might
be to run `(asdf:compile-system :advent :force t)` or similar inside
AOC::BUILD-SITE but I hate that solution so much I can hardly speak.
I just don't want to think about it anymore tonight, so I'll sleep on it.