Minor fix to Day 10 formatting for the site build.
Previously, the answer to Day 10 overflowed the summary box and,
because that output is parsed as markdown by 3bmd, it created additional
headings/sections in the finished page. Now DEFSUMMARY is more flexible so
we can just omit the answers in cases where they don't format nicely.

Sure, it would be nice to keep everything inside the quote block but
I'll worry about that another day.